Removed custom deserialization, moved field validation into the upload schema validation, and moved validator tests to their own test classes
