+++ app/org/sagebionetworks/bridge/play/controllers/UploadSchemaController.java
@@ -9,7 +9,6 @@ import org.springframework.stereotype.Controller;
 
 import play.mvc.Result;
 
-import org.sagebionetworks.bridge.dynamodb.DynamoUploadSchema;
 import org.sagebionetworks.bridge.models.accounts.UserSession;
 import org.sagebionetworks.bridge.models.studies.StudyIdentifier;
 import org.sagebionetworks.bridge.models.upload.UploadSchema;
@@ -37,7 +36,7 @@ public class UploadSchemaController extends BaseController {
         UserSession session = getAuthenticatedSession(DEVELOPER);
         StudyIdentifier studyId = session.getStudyIdentifier();
         
-        UploadSchema uploadSchema = DynamoUploadSchema.fromJson(requestToJSON(request()));
+        UploadSchema uploadSchema = parseJson(request(), UploadSchema.class);
         UploadSchema createdSchema = uploadSchemaService.createOrUpdateUploadSchema(studyId, uploadSchema);
         return okResult(createdSchema);
     }
