+++ app/org/sagebionetworks/bridge/services/StudyServiceImpl.java
@@ -178,6 +178,7 @@ public class StudyServiceImpl implements StudyService {
     public Study updateStudy(Study study, boolean isAdminUpdate) {
         checkNotNull(study, Validate.CANNOT_BE_NULL, "study");
         
+        setDefaultsIfAbsent(study); 
         sanitizeHTML(study);
 
         // These cannot be set through the API and will be null here, so they are set on update
@@ -291,17 +292,12 @@ public class StudyServiceImpl implements StudyService {
     }
     
     private EmailTemplate sanitizeEmailTemplate(EmailTemplate template) {
-        String subject = template.getSubject();
-        if (StringUtils.isNotBlank(subject)) {
-            subject = Jsoup.clean(subject, Whitelist.none());    
-        }
-        String body = template.getBody();
-        if (StringUtils.isNotBlank(body)) {
-            if (template.getMimeType() == MimeType.TEXT) {
-                body = Jsoup.clean(template.getBody(), Whitelist.none());
-            } else {
-                body = Jsoup.clean(template.getBody(), Whitelist.relaxed());
-            }
+        String subject = Jsoup.clean(template.getSubject(), Whitelist.none());
+        String body = null;
+        if (template.getMimeType() == MimeType.TEXT) {
+            body = Jsoup.clean(template.getBody(), Whitelist.none());
+        } else {
+            body = Jsoup.clean(template.getBody(), Whitelist.relaxed());
         }
         return new EmailTemplate(subject, body, template.getMimeType());
     }
