+++ test/org/sagebionetworks/bridge/models/upload/UploadSchemaTest.java
@@ -4,21 +4,14 @@ import static org.junit.Assert.assertEquals;
 import static org.junit.Assert.assertFalse;
 import static org.junit.Assert.assertTrue;
 
-import java.util.ArrayList;
-import java.util.Collections;
-import java.util.HashMap;
 import java.util.List;
 import java.util.Map;
 
 import org.junit.Test;
-import org.sagebionetworks.bridge.dynamodb.DynamoUploadFieldDefinition;
 import org.sagebionetworks.bridge.dynamodb.DynamoUploadSchema;
 import org.sagebionetworks.bridge.exceptions.InvalidEntityException;
 import org.sagebionetworks.bridge.json.BridgeObjectMapper;
 import org.sagebionetworks.bridge.json.JsonUtils;
-import org.sagebionetworks.bridge.validators.UploadSchemaValidator;
-import org.sagebionetworks.bridge.validators.Validate;
-import org.springframework.validation.MapBindingResult;
 
 @SuppressWarnings("unchecked")
 public class UploadSchemaTest {
@@ -112,222 +105,6 @@ public class UploadSchemaTest {
         ddbUploadSchema.setKey("api:test:schema");
         assertEquals("api", ddbUploadSchema.getStudyId());
         assertEquals("test:schema", ddbUploadSchema.getSchemaId());
-    }
-
-    // branch coverage
-    @Test
-    public void validatorSupportsClass() {
-        assertTrue(UploadSchemaValidator.INSTANCE.supports(UploadSchema.class));
-    }
-
-    // branch coverage
-    @Test
-    public void validatorSupportsSubclass() {
-        assertTrue(UploadSchemaValidator.INSTANCE.supports(DynamoUploadSchema.class));
-    }
-
-    // branch coverage
-    @Test
-    public void validatorDoesntSupportClass() {
-        assertFalse(UploadSchemaValidator.INSTANCE.supports(String.class));
-    }
-
-    // branch coverage
-    // we call the validator directly, since Validate.validateThrowingException filters out nulls and wrong types
-    @Test
-    public void validateNull() {
-        MapBindingResult errors = new MapBindingResult(new HashMap<>(), "UploadSchema");
-        UploadSchemaValidator.INSTANCE.validate(null, errors);
-        assertTrue(errors.hasErrors());
-    }
-
-    // branch coverage
-    // we call the validator directly, since Validate.validateThrowingException filters out nulls and wrong types
-    @Test
-    public void validateWrongClass() {
-        MapBindingResult errors = new MapBindingResult(new HashMap<>(), "UploadSchema");
-        UploadSchemaValidator.INSTANCE.validate("this is the wrong class", errors);
-        assertTrue(errors.hasErrors());
-    }
-
-    @Test
-    public void validateHappyCase() {
-        // set up schema to validate
-        DynamoUploadSchema schema = new DynamoUploadSchema();
-        schema.setName("happy schema");
-        schema.setSchemaId("happy-schema");
-        schema.setStudyId("test-study");
-        schema.setSchemaType(UploadSchemaType.IOS_DATA);
-
-        // test field def list
-        List<UploadFieldDefinition> fieldDefList = new ArrayList<>();
-        fieldDefList.add(new DynamoUploadFieldDefinition.Builder().withName("test-field")
-                .withType(UploadFieldType.ATTACHMENT_BLOB).build());
-        schema.setFieldDefinitions(fieldDefList);
-
-        // validate
-        Validate.entityThrowingException(UploadSchemaValidator.INSTANCE, schema);
-    }
-
-    @Test
-    public void validateHappyCase2() {
-        // set up schema to validate
-        DynamoUploadSchema schema = new DynamoUploadSchema();
-        schema.setName("happy schema 2");
-        schema.setRevision(1);
-        schema.setSchemaId("happy-schema-2");
-        schema.setStudyId("test-study");
-        schema.setSchemaType(UploadSchemaType.IOS_SURVEY);
-
-        // test field def list
-        List<UploadFieldDefinition> fieldDefList = new ArrayList<>();
-        fieldDefList.add(new DynamoUploadFieldDefinition.Builder().withName("foo-field")
-                .withType(UploadFieldType.INT).build());
-        fieldDefList.add(new DynamoUploadFieldDefinition.Builder().withName("bar-field")
-                .withType(UploadFieldType.STRING).build());
-        schema.setFieldDefinitions(fieldDefList);
-
-        // validate
-        Validate.entityThrowingException(UploadSchemaValidator.INSTANCE, schema);
-    }
-
-    @Test(expected = InvalidEntityException.class)
-    public void validateNullFieldDefList() {
-        // set up schema to validate
-        DynamoUploadSchema schema = new DynamoUploadSchema();
-        schema.setName("test schema");
-        schema.setSchemaId("test-schema");
-        schema.setStudyId("test-study");
-        schema.setSchemaType(UploadSchemaType.IOS_DATA);
-
-        // validate
-        Validate.entityThrowingException(UploadSchemaValidator.INSTANCE, schema);
-    }
-
-    @Test(expected = InvalidEntityException.class)
-    public void validateEmptyFieldDefList() {
-        // set up schema to validate
-        DynamoUploadSchema schema = new DynamoUploadSchema();
-        schema.setName("test schema");
-        schema.setSchemaId("test-schema");
-        schema.setStudyId("test-study");
-        schema.setFieldDefinitions(Collections.<UploadFieldDefinition>emptyList());
-        schema.setSchemaType(UploadSchemaType.IOS_DATA);
-
-        // validate
-        Validate.entityThrowingException(UploadSchemaValidator.INSTANCE, schema);
-    }
-
-    @Test(expected = InvalidEntityException.class)
-    public void validateNullName() {
-        // set up schema to validate
-        DynamoUploadSchema schema = new DynamoUploadSchema();
-        schema.setSchemaId("test-schema");
-        schema.setStudyId("test-study");
-        schema.setSchemaType(UploadSchemaType.IOS_DATA);
-
-        // test field def list
-        List<UploadFieldDefinition> fieldDefList = new ArrayList<>();
-        fieldDefList.add(new DynamoUploadFieldDefinition.Builder().withName("good-field")
-                .withType(UploadFieldType.ATTACHMENT_BLOB).build());
-        schema.setFieldDefinitions(fieldDefList);
-
-        // validate
-        Validate.entityThrowingException(UploadSchemaValidator.INSTANCE, schema);
-    }
-
-    @Test(expected = InvalidEntityException.class)
-    public void validateEmptyName() {
-        // set up schema to validate
-        DynamoUploadSchema schema = new DynamoUploadSchema();
-        schema.setName("");
-        schema.setSchemaId("test-schema");
-        schema.setStudyId("test-study");
-        schema.setSchemaType(UploadSchemaType.IOS_DATA);
-
-        // test field def list
-        List<UploadFieldDefinition> fieldDefList = new ArrayList<>();
-        fieldDefList.add(new DynamoUploadFieldDefinition.Builder().withName("good-field")
-                .withType(UploadFieldType.ATTACHMENT_BLOB).build());
-        schema.setFieldDefinitions(fieldDefList);
-
-        // validate
-        Validate.entityThrowingException(UploadSchemaValidator.INSTANCE, schema);
-    }
-
-    @Test(expected = InvalidEntityException.class)
-    public void validateNegativeRev() {
-        // set up schema to validate
-        DynamoUploadSchema schema = new DynamoUploadSchema();
-        schema.setName("test schema");
-        schema.setRevision(-1);
-        schema.setSchemaId("test-schema");
-        schema.setStudyId("test-study");
-        schema.setSchemaType(UploadSchemaType.IOS_DATA);
-
-        // test field def list
-        List<UploadFieldDefinition> fieldDefList = new ArrayList<>();
-        fieldDefList.add(new DynamoUploadFieldDefinition.Builder().withName("good-field")
-                .withType(UploadFieldType.ATTACHMENT_BLOB).build());
-        schema.setFieldDefinitions(fieldDefList);
-
-        // validate
-        Validate.entityThrowingException(UploadSchemaValidator.INSTANCE, schema);
-    }
-
-    @Test(expected = InvalidEntityException.class)
-    public void validateNullSchemaId() {
-        // set up schema to validate
-        DynamoUploadSchema schema = new DynamoUploadSchema();
-        schema.setName("test schema");
-        schema.setStudyId("test-study");
-        schema.setSchemaType(UploadSchemaType.IOS_DATA);
-
-        // test field def list
-        List<UploadFieldDefinition> fieldDefList = new ArrayList<>();
-        fieldDefList.add(new DynamoUploadFieldDefinition.Builder().withName("good-field")
-                .withType(UploadFieldType.ATTACHMENT_BLOB).build());
-        schema.setFieldDefinitions(fieldDefList);
-
-        // validate
-        Validate.entityThrowingException(UploadSchemaValidator.INSTANCE, schema);
-    }
-
-    @Test(expected = InvalidEntityException.class)
-    public void validateEmptySchemaId() {
-        // set up schema to validate
-        DynamoUploadSchema schema = new DynamoUploadSchema();
-        schema.setName("test schema");
-        schema.setSchemaId("");
-        schema.setStudyId("test-study");
-        schema.setSchemaType(UploadSchemaType.IOS_DATA);
-
-        // test field def list
-        List<UploadFieldDefinition> fieldDefList = new ArrayList<>();
-        fieldDefList.add(new DynamoUploadFieldDefinition.Builder().withName("good-field")
-                .withType(UploadFieldType.ATTACHMENT_BLOB).build());
-        schema.setFieldDefinitions(fieldDefList);
-
-        // validate
-        Validate.entityThrowingException(UploadSchemaValidator.INSTANCE, schema);
-    }
-
-    @Test(expected = InvalidEntityException.class)
-    public void validateNullSchemaType() {
-        // set up schema to validate
-        DynamoUploadSchema schema = new DynamoUploadSchema();
-        schema.setName("test schema");
-        schema.setSchemaId("test-schema");
-        schema.setStudyId("test-study");
-
-        // test field def list
-        List<UploadFieldDefinition> fieldDefList = new ArrayList<>();
-        fieldDefList.add(new DynamoUploadFieldDefinition.Builder().withName("test-field")
-                .withType(UploadFieldType.ATTACHMENT_BLOB).build());
-        schema.setFieldDefinitions(fieldDefList);
-
-        // validate
-        Validate.entityThrowingException(UploadSchemaValidator.INSTANCE, schema);
     }
 
     @Test
